Fix types used in ethercat

+++ march_hardware/src/EtherCAT/EthercatSDO.cpp
@@ -4,18 +4,11 @@
 
 #include <soem/ethercattype.h>
 #include <soem/nicdrv.h>
-#include <soem/ethercatbase.h>
-#include <soem/ethercatmain.h>
-#include <soem/ethercatdc.h>
 #include <soem/ethercatcoe.h>
-#include <soem/ethercatfoe.h>
-#include <soem/ethercatconfig.h>
-#include <soem/ethercatprint.h>
 
 namespace march
 {
-// TODO(Isha, Martijn, Tim) refactor this with more generic types
-int sdo_bit8(int slave, uint32_t index, uint8_t sub, uint8_t value)
+int sdo_bit8(uint16_t slave, uint16_t index, uint8_t sub, uint8_t value)
 {
   ROS_DEBUG("sdo_bit8: slaveIndex %i, reg 0x%X, subindex %i, value 0x%X", slave, index, sub, value);
   int received_working_counter = ec_SDOwrite(slave, index, sub, FALSE, 1, &value, EC_TIMEOUTRXM);
@@ -27,7 +20,7 @@ int sdo_bit8(int slave, uint32_t index, uint8_t sub, uint8_t value)
   return received_working_counter;
 }
 
-int sdo_bit16(int slave, uint32_t index, uint8_t sub, uint16_t value)
+int sdo_bit16(uint16_t slave, uint16_t index, uint8_t sub, uint16_t value)
 {
   ROS_DEBUG("sdo_bit16: slaveIndex %i, reg 0x%X, subindex %i, value 0x%X", slave, index, sub, value);
   int received_working_counter = ec_SDOwrite(slave, index, sub, FALSE, 2, &value, EC_TIMEOUTRXM);
@@ -39,7 +32,7 @@ int sdo_bit16(int slave, uint32_t index, uint8_t sub, uint16_t value)
   return received_working_counter;
 }
 
-int sdo_bit32(int slave, uint32_t index, uint8_t sub, uint32_t value)
+int sdo_bit32(uint16_t slave, uint16_t index, uint8_t sub, uint32_t value)
 {
   ROS_DEBUG("sdo_bit32: slaveIndex %i, reg 0x%X, subindex %i, value 0x%X", slave, index, sub, value);
   int received_working_counter = ec_SDOwrite(slave, index, sub, FALSE, 4, &value, EC_TIMEOUTRXM);
